requirements.txt
Django
djangorestframework==3.14.0
django-cors-headers==4.3.1
djoser==2.2.2
# psycopg2-binary==2.9.9
Pillow==10.1.0
django-filter==23.5
django-extensions==3.2.3
django-storages==1.14.2
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.1
drf-yasg==1.21.7
djangorestframework-gis==1.0
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
